confirmations in adminhomepage and cell selection

+++ AdminHomePage.vb
@@ -4,12 +4,14 @@ Imports System.Data.SqlClient
 'Admin Home Page
 
 Public Class AdminHomePage
+    'logout user
     Private Sub LogoutAdmin_Click(sender As Object, e As EventArgs) Handles LogoutAdmin.Click
         LoginForm.sql.Close()
-        LoginForm.Show()
+        'LoginForm.Show()
         Me.Close()
     End Sub
 
+    'when page loads
     Private Sub AdminHomePage_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         Upd_Name.Enabled = False
         Upd_Designation.Enabled = False
@@ -46,30 +48,41 @@ Public Class AdminHomePage
     End Sub
 
     Private Sub AddRow_Click(sender As Object, e As EventArgs) Handles AddRow.Click
+
         If Add_ID.Text = Nothing Or Add_EmpName.Text = Nothing Or Add_Designtion.Text = Nothing Or Add_Password.Text = Nothing Then
             MsgBox("Please fill all the details!")
         Else
+
+            'if designation is employee(E) or manager(M)
             If Add_Designtion.Text = "E" Or Add_Designtion.Text = "M" Then
-                Dim AddCommand As String = "INSERT INTO EMPLOYEES (Id,empName,Designation,password) VALUES ('" + Add_ID.Text + "','" + Add_EmpName.Text + "','" + Add_Designtion.Text + "','" + Add_Password.Text + "')"
 
-                'creating a sql command statement 
-                Dim command As SqlCommand = LoginForm.sql.CreateCommand()
-                command.CommandText = AddCommand
+                'conformation
+                Dim responce As String = MsgBox("Add '" & Add_ID.Text & "' to the database", vbYesNo, "Add User?")
+                If responce = vbYes Then
 
-                'sqladapter to handle the sql commands 
-                Dim sqlAdapter As New SqlDataAdapter With {
-                .SelectCommand = command
-            }
-                'creates a table with the required data
-                Dim data As New DataSet()
-                sqlAdapter.Fill(data)
+                    'adds data to the table
+                    Dim AddCommand As String = "INSERT INTO EMPLOYEES (Id,empName,Designation,password) VALUES ('" + Add_ID.Text + "','" + Add_EmpName.Text + "','" + Add_Designtion.Text + "','" + Add_Password.Text + "')"
 
-                DataLoader()
+                    'creating a sql command statement 
+                    Dim command As SqlCommand = LoginForm.sql.CreateCommand()
+                    command.CommandText = AddCommand
 
-                Add_ID.Text = Nothing
-                Add_EmpName.Text = Nothing
-                Add_Designtion.Text = Nothing
-                Add_Password.Text = Nothing
+                    'sqladapter to handle the sql commands 
+                    Dim sqlAdapter As New SqlDataAdapter With {
+                    .SelectCommand = command
+                }
+                    'creates a table with the required data
+                    Dim data As New DataSet()
+                    sqlAdapter.Fill(data)
+
+                    DataLoader()
+
+                    'empty the textboxes
+                    Add_ID.Text = Nothing
+                    Add_EmpName.Text = Nothing
+                    Add_Designtion.Text = Nothing
+                    Add_Password.Text = Nothing
+                End If
             Else
                 MsgBox("Invalid Designation!")
             End If
@@ -77,14 +90,17 @@ Public Class AdminHomePage
     End Sub
 
     Private Sub DataLoader()
+
         'creating a sql command statement 
         Dim Consolecommand As SqlCommand = LoginForm.sql.CreateCommand()
-        Consolecommand.CommandText = "SELECT * FROM Employees WHERE designation <> 'A'"
+        Consolecommand.CommandText = "SELECT * FROM Employees WHERE Id <> '" & AdminId.Text & "'"
+        'Consolecommand.CommandText = "SELECT * FROM Employees WHERE designation <> 'A'"
 
         'sqladapter to handle the sql commands 
         Dim ConsolesqlAdapter As New SqlDataAdapter With {
             .SelectCommand = Consolecommand
         }
+
         'creates a table with the required data
         Dim Consoledata As New DataSet()
         ConsolesqlAdapter.Fill(Consoledata)
@@ -96,31 +112,35 @@ Public Class AdminHomePage
         If Del_ID.Text = Nothing Then
             MsgBox("Please mention the UserID ")
         Else
+            'conformation
+            Dim responce As String = MsgBox("Delete '" & Del_ID.Text & "' from the database", vbYesNo, "Delete User?")
+            If responce = vbYes Then
+                Dim DelCommand As String = "DELETE FROM Employees WHERE Id ='" + Del_ID.Text + "'"
 
-            Dim DelCommand As String = "DELETE FROM Employees WHERE Id ='" + Del_ID.Text + "'"
+                'creating a sql command statement 
+                Dim command As SqlCommand = LoginForm.sql.CreateCommand()
+                command.CommandText = "select * from Employees where Id ='" + Del_ID.Text + "'"
 
-            'creating a sql command statement 
-            Dim command As SqlCommand = LoginForm.sql.CreateCommand()
-            command.CommandText = "select * from Employees where Id ='" + Del_ID.Text + "'"
-
-            'sqladapter to handle the sql commands 
-            Dim sqlAdapter As New SqlDataAdapter With {
+                'sqladapter to handle the sql commands 
+                Dim sqlAdapter As New SqlDataAdapter With {
                 .SelectCommand = command
-            }
-            'creates a table with the required data
-            Dim data As New DataSet()
-            sqlAdapter.Fill(data)
+                }
 
-            If (data.Tables(0).Rows.Count) > 0 Then
-                command.CommandText = DelCommand
-                sqlAdapter.SelectCommand = command
+                'creates a table with the required data
+                Dim data As New DataSet()
                 sqlAdapter.Fill(data)
-                DataLoader()
-            Else
-                MsgBox("User not found!")
-            End If
-            Del_ID.Text = Nothing
 
+                If (data.Tables(0).Rows.Count) > 0 Then
+                    command.CommandText = DelCommand
+                    sqlAdapter.SelectCommand = command
+                    sqlAdapter.Fill(data)
+                    DataLoader()
+                Else
+                    MsgBox("User not found!")
+                End If
+                'reseting textbox
+                Del_ID.Text = Nothing
+            End If
         End If
     End Sub
 
@@ -130,31 +150,37 @@ Public Class AdminHomePage
         Else
             If Upd_Designation.Text = "E" Or Upd_Designation.Text = "M" Then
 
-                Dim UpdateCommand As String = "UPDATE Employees SET empName ='" + Upd_Name.Text + "',Designation ='" + Upd_Designation.Text + "',password='" + Upd_Password.Text + "' Where Id ='" + Upd_ID.Text + "'"
-                'creating a sql command statement 
-                Dim command As SqlCommand = LoginForm.sql.CreateCommand()
-                command.CommandText = UpdateCommand
+                'conformation
+                Dim responce As String = MsgBox("Update User?", vbYesNo, "Update " & Upd_ID.Text & " in the database")
+                If responce = vbYes Then
 
-                'sqladapter to handle the sql commands 
-                Dim sqlAdapter As New SqlDataAdapter With {
+                    Dim UpdateCommand As String = "UPDATE Employees SET empName ='" + Upd_Name.Text + "',Designation ='" + Upd_Designation.Text + "',password='" + Upd_Password.Text + "' Where Id ='" + Upd_ID.Text + "'"
+                    'creating a sql command statement 
+                    Dim command As SqlCommand = LoginForm.sql.CreateCommand()
+                    command.CommandText = UpdateCommand
+
+                    'sqladapter to handle the sql commands 
+                    Dim sqlAdapter As New SqlDataAdapter With {
                     .SelectCommand = command
-                }
-                'creates a table with the required data
-                Dim data As New DataSet()
-                sqlAdapter.Fill(data)
+                    }
 
-                DataLoader()
+                    'creates a table with the required data
+                    Dim data As New DataSet()
+                    sqlAdapter.Fill(data)
 
-                Upd_Name.Text = Nothing
-                Upd_Designation.Text = Nothing
-                Upd_Password.Text = Nothing
-                Upd_ID.Text = Nothing
+                    DataLoader()
 
-                Upd_Name.Enabled = False
-                Upd_Designation.Enabled = False
-                Upd_Password.Enabled = False
-                UpdateRow.Enabled = False
+                    'reseting textbox
+                    Upd_Name.Text = Nothing
+                    Upd_Designation.Text = Nothing
+                    Upd_Password.Text = Nothing
+                    Upd_ID.Text = Nothing
 
+                    Upd_Name.Enabled = False
+                    Upd_Designation.Enabled = False
+                    Upd_Password.Enabled = False
+                    UpdateRow.Enabled = False
+                End If
             Else
                 MsgBox("Invalid Designation!")
             End If
@@ -190,5 +216,18 @@ Public Class AdminHomePage
                 MsgBox("User not found!")
             End If
         End If
+    End Sub
+
+    Sub itemclicked()
+
+    End Sub
+
+    Private Sub AdminDataGrid_CellContentClick(sender As Object, e As DataGridViewCellEventArgs) Handles AdminDataGrid.CellContentClick
+        'itemclicked()
+        With AdminDataGrid
+            Dim i As Integer = .CurrentCell.RowIndex
+            Upd_ID.Text = .Rows(i).Cells(0).Value.ToString
+            Del_ID.Text = .Rows(i).Cells(0).Value.ToString
+        End With
     End Sub
 End Class
